refactor: resurrected teledisko task management
