+++ src/state/neokingdom/odoo/index.js
@@ -1,3 +1,1 @@
-export * from "./agent";
-export * from "./user";
 export * from "./project";
