+++ src/state/odoo/index.js
@@ -1,0 +1,2 @@
+export * from "./agent";
+export * from "./user";
